Reworked Core Mod Player Logic
There was some core issues with how the logic for the module player
worked.  After some tests against more advanced Mod files, more are
working with advanced features.  There a few that has some weird edge
case logic that'll need to be addressed.  The core test modules still
work, but need close examination to ensure they are being played
correctly.  Currently using MilkyTracker to do comparisons.

More debug options will be needed to track the logic behind the scenes
and see if there's some strange edge cases happening.

Also found a issue with the repeat looping feature that meant it was not
working at all.

+++ StrangeGameEngine/include/SGE/sound.h
@@ -295,10 +295,10 @@ namespace SGE
 			int Size = 0;
 			
 			//  The offeset the sample will repeat at.
-			unsigned int RepeatOffset = 0;
+			float RepeatOffset = 0;
 
 			//  The duration of the repeat
-			unsigned int RepeatDuration = 0;
+			float RepeatDuration = 0;
 
 			//  Create a blank buffer of a certain sample size
 			void SGEAPI Create(int numOfSamples);
